auth/requirements.txt
asgiref==3.6.0
astroid==2.12.13
async-timeout==4.0.2
autopep8==2.0.1
dill==0.3.6
Django==4.1.4
django-redis==5.2.0
djangorestframework==3.14.0
gunicorn==20.1.0
isort==5.11.3
lazy-object-proxy==1.8.0
mccabe==0.7.0
mysqlclient==2.1.1
platformdirs==2.6.0
pycodestyle==2.10.0
PyJWT==2.6.0
pylint==2.15.9
pytz==2022.7
redis==4.4.0
sqlparse==0.4.3
tomli==2.0.1
tomlkit==0.11.6
wrapt==1.14.1
